libintelprov: fixes related 64bit build and dependencies
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7119
Reviewed-on: https://android.intel.com:443/291375

+++ flash.c
@@ -43,7 +43,7 @@ int flash_custom_boot(void *data, size_t size)
 	return ops_call(ifwi, flash_custom_boot, data, size);
 }
 
-int flash_token_umip(void *data, size_t size)
+int flash_token_umip(void *data, unsigned size)
 {
 	return ops_call(ifwi, flash_token_umip, data, size);
 }
